Fix span issues
Fix the padding/visual offset of the highlight (i.e. the jagged
underline) in diagnostic message not showing up.
Fix the lexer not creating the correct span when reporting illegal
characters (solves crashes on non-ASCII Unicode input)
Fix the length of the highlight not being computed correctly when
resolving spans (solves crashes on non-ASCII Unicode input)
Fix the lexer not respecting new newly updated span system (where the
first source file now starts at 1 as opposed to 0).
Update tests for the lexer to accommodate the fix above.
Enhance the test suite of the lexer by requiring to now also test the
spans reported by diagnostic messages. This is a regression test for
the issue one below the top.