Make many desugaring errors non-fatal
Introduce the concept of invalid nodes which are used for non-fatal
error reporting. This special node does not store any information, the
errors are collected separately into a `Bag<Diagnostic>`.
We use this technique to non-fatalize the error "missing mandatory type
annotation".
Note that there is a second way of reporting non-fatal errors:
`AccumulateNErrors`. This way is prefered over the previously discussed
when the data being processed is truly independent from each other.
(`TryNonFatally` the fake node as fake data to resolve the dependency).