Improve printing of the HIR
Let's just say this fixes #12 even though we should do much better in
the future like "intelligent" indentation, better path printing
(omitting `crate` if shadowing does not conflict), etcetera. This is a
MVP which respects syntactic precedence.

We also add a ton of tests and a VS Code highlighting for our grammar
file.