Work on laziness marker
You can now ascribe the keyword `lazy` to function parameters which
makes the argument in the application of the function implicitly coerce
to thunk, that is into a function from unit to the original type.
The design is not good right now, it is too boilerplate-y and the naming
is just wrong. I gonna write digitalize my notes in the issue tracker I
guess reflecting on this implementation.

Also "funnily", what I've come to know, all applications are somewhat
lazy right now: If the (formal) parameter is not used than the (actual)
parameter/argument never gets evaluated even though by specification (in
my head), the language should be strict (with optional laziness, this
feature) and also not Church-Rosser. I strongly believe this bug is due
to explicit subsitutions which are inherently lazy obviously.

So although `lazy` seems to work, the rest, the non-lazy part, is wrong.
We gonna fix that later, it's not a priority.