Properly report errors on undefined exposure reach
Previously, such reaches were only reported if the binding it restricted
in exposure was actually used/referenced. Further, we do not crash
anymore if the reach is the same binding as the binding we annotate or
if it's a descendant of it. Just report an error. Additionally verify
that the reach is a module.