Change the delimiter of data and module decls
From the reserved punctuation `=` to the keyword `of`.
This is preparation for the new indentation system which allows for more
flexible and liberal indentation and which should ideally be based on
solely a few simple rules (in the lexer). Therefore, we need to
streamline the syntax and make it more predictable for the lexer.
More concretely, with this change, we can finally state that after an `=`
we do not necessarily indent the content that comes after it.
Previously, this was not the case: The equals sign in module and data declarations
required indentation. As a replacement, `of` is a perfect choice: It
already requires indentation after it (in case analyses) and it's
available for and readable in data/module declarations. Everything looks
so much better now!