Fix decl. indices of namespaces b. misinterpreted
They used to be local declaration indices but in other parts of the code
we did not respect that they are local relative to the namespace not the
to current crate (since we did not adjust the code from pre-crate-system).
Now, we simply store global declaration indices in a namespace which is
way simpler. I also tried keeping the local indices but then defined an
accessor method resolving them to global indices given three more
arguments (!). It was ugly and error-prone.