Remove the limit of 64 on number of attributes
We used to use a bitset for performant filtering (neglegible difference though probably).
The bitset was called AttributeKeys. Instead, we now have a new trait Query (basically a
predicate on AttributeKinds) where the most prominent implementor is the type AttributeName
which is derived from AttributeKind via the proc macro Discriminant. Maybe less performant
and less space efficient but at least it does not restrict the total number of attributes
that can be added to the language.

Also, we started working on displaying attributes nicely in generated documentations.