Start implementing declaration links in documentation comments
We register an AsciiDoc macro `decl` in Ruby which uses a custom
file-based IPC protocol to relay the data back to the compiler.
It's quite elaborate what we have to do to successfully bridge
Asciidoctor (written in Ruby) and Rust (file watcher in a separate thread,
a lot of shared variables, 4 different files, serializing and
deserializing, etcetera).

However, the integration with Asciidoctor is not sustainable, it is not
performant at all! There is way too much overhead. It's not the file watchers
(neither the Rust nor the Ruby one) which continuously watch/probe certain
files every 100 or 200ms. It's the overhead to execute Asciidoctor as a
child process ... for every single documented declaration anew!

Documenting `core` (with its idk 18 documented declarations) takes a
whopping 5.55 seconds! Even if we comment out the file watchers.

Executing Asciidoctor for a single-line documentation, we have to wait
255 seconds on average for it to finish! According to Asciidoctor and
to the shell built-in time command, the "user time" (the time of the
actual logic) is 0.01s but the total time (execution overhead) is
0.265s (just executing `asciidoctor` in the terminal w/o Lushui)!

Of course one might say it's obvious that we shouldn't call it so often,
that we should collect/buffer the documentation of everys single declaration,
concatenate it with some kind of separator and then execute AsciiDoctor
once (per crate). But this is just so annoying: How do I communicate
`url_prefix` (basically the relative path to the current module)
to the `decl` macro? It may vary from declaration to declaration.
I would need to do preprocessing I guess and I'd have to re-architect the
way the documenter works (well, I will have to do so anyway at some point
for concurrent documentation generation) and what's more I don't like
it that we have a dependency on a Ruby program. I wan't the compiler
to be "self-containted".

Only solution: A custom Rust implementation for AsciiDoc!