Make it so out-of-line modules in root modules have to reside in subfolders by default
Fixes #129.

Previously, all out-of-line modules (without a `@location` attribute) had to be placed in a subfolder
*except for* out-of-line modules in root modules. It's a confusing inconsistency which I did not really
notice since I blindly copied it from Rust (edition >= 2018). A particular nice advantage is that
files from separate components (e.g. a library and an executable) no longer "overlap"/reside in the same
folder by default.

I added a lot of tests for out-of-line modules and adjusted the standard library `core` to work with the new
system.

Finally, I noticed that we'd previously always set the name/binder of the root module to the component name
instead of the filename unlike for out-of-line modules which is incorrect for complex packages:
Consider the files `library.lushui` and `main.lushui` whose root module would be bound to `pkg` if the
package name was `pkg`. This is obviously nonsense and had to be fixed for the new system to work:
File `library.lushui` expects its out-of-line modules to reside in the subfolder `library/` by default.