Move reporter and source map into build session and some small refactorings
Turn the BuildSession further into a “god object” by storing the session reporter
and session source map in it and changing almost every pass etc. to use those two
via the build session struct. This takes away the flexibility for users of those
passes to specify a map and reporter per invocation obviously.
On the other, in 95% of cases, that's what we want and it reduces a lot of boilerplate/
parameter passing.

Not being able to choose a reporter is actually a problem: Word::parse uses the lexer
under the hood and does not want any diagnostics reported. That's why the lexer still takes
a reporter parameter.