Temporarily get rid of staticvec dependency
It was a pain to maintain without any gain (^^).
It constantly broke since it depends on unstable features from a nightly rustc for
advanced const generics and const evaluation functionality.
And it's only ever used in the abortive / atrophied / stunted compiler::compiler
(codegen) module which I added years ago. Once we start working on the massive
feature of code generation to a custom VM (?) again, we can potentially re-add
staticvec. Until now it lead to nothing but sorrow.