Decouple components from packages and program entries
Components no longer need to be part of a package.
Single source files are no longer wrapped up in a dummy package.
The nonsensical concept of “single-file packages” is now gone.

Components no longer store the potential program entry (`main`).
Instead, the resolver now offers a method that looks it up.
This makes the concept of components simpler and more light-weight.
A program entry is part of a different layer: The layer of the
“driver” (e.g. `main.rs`).
Currently, we still store the “component type” in the component.
However, this also violates the layer model: Component types are
actually part of the package system and are very high-level:
Benchmark suites, test suites etc should not be coupled so tightly
to the concept of components. In the future, that field needs to go,
too.