LLVM backend: Compile definitely-total non-lambda expressions to global variables
Basically avoiding “thunking” / “delaying” certain basic expressions that we know
are definitely total and pure. Not sure if this is premature optimization but I
hope that this extra step might simplify the life of LLVM.

Apart from the that, handle out of order / nonlinear declarations by having two
codegen passes.