Do not report any errors in the lexer and do not bail out early
Instead, defer error reporting to the parser. This makes the lexer more light-weight, versatile and flexible.

The plan is to use it in the golden testing framework to properly parse comments
(for test configurations) which enables it to use Spans for error reporting.
Before, lexing could fail (Result) which would render the testing framework useless
if it wanted to test exactly such lexing errors.

Further, it now wouldn't need to pass a silent reporter to the lexer just to
retrieve comments meaning the lexer now has a cleaner API.

The lexer has become more lenient: Trailing and consecutive dashes in identifiers are
now allowed. As well as digits immediately following a dash inside of an identifier.
Additionally, trailing and consecutive apostrophes in number literals are now allowed.