Allow closing brackets to be further dedented
Previously, we would insert a line break before such a closing brackets
according to the rules. However, I'd like to be able to place my closing
bracket at the same indentation level as the respective opening one.

See the updated UI tests for example.

The current approach is quite liberal and careless. In the future we
should probably restrict this a bit, e.g. forbidding closing brackets
to be further indented than the respective opening one.